Add initial support for X20
* testing

* Update eventbuild.yml

* Update x20

* x20-stuff

* Update 00-run-chroot.sh

* Update 00-run-chroot.sh

* Update 00-run-chroot.sh

* cleanup

* cleanup

* now everything should start normally

* Update 00-run-chroot.sh

* Update 00-run-chroot.sh

* change name

* give him a little sleep time

* Update 00-run-chroot.sh

* Update 00-run-chroot.sh

* Update 00-run-chroot.sh

* Update 00-run-chroot.sh

* Update 00-run-chroot.sh

* Update eventbuild.yml

* Update eventbuild.yml

* remove stuff

* remove graphic drivers

* shrink the image a little

* let's see if pishrink works

* ok .. looks like this is a no

* Update 00-run-chroot.sh

* Update 00-run-chroot.sh

* Update eventbuild.yml

+++ stages/01-Baseimage/01-run.sh
@@ -15,7 +15,7 @@ pushd ${STAGE_WORK_DIR}
     echo "bytes"
     ls -l
 
-if [[ "${OS}" != ubuntu-x86 ]]; then
+if [[ "${OS}" != ubuntu-x86 ]] && [[ "${OS}" != debian-X20 ]]; then
 
     log "Create empty image" #this will be attached to the base image to increase the size of it
     dd if=/dev/zero of=temp.img bs=1 count=1 seek=$DIFFERENCE
